Add files via upload
First Knit
jesus avalos
2023-06-27
1 A Little Survey
1.1 A short history of you.
1.1.1 my answer
1.2 Your computer and your and its capability?
1.2.1 my answer
1.3 Others and You?
1.3.1 my answer
1.4 Aspirations?
1.4.1 my answer
1.5 Past adventures?
1.5.1 my answer
1.6 Why are you taking Statistics?
1.6.1 my answer
1 A Little Survey
Here are just a few questions so I can get to know you and so you can get practice submitting an assignment in Canvas. For each question, insert your answer underneath. Make sure each answer is at the minimum a complete sentence with correctly spelled words and correct puncuation.

1.1 A short history of you.
Where did you attend High School, and did you graduate? Compare english class to math, is one easier for you than the other? Finally which subject did you do the best in? Your answer should have at least two sentences devoted to each question.

1.1.1 my answer
I went to Long Beach Polytechnic High school,I graduated in 2002. I found Math a bit easier than english. I did have the same grades in both subjects. I did very well in my wood shop and auto mechanics class.

1.2 Your computer and your and its capability?
For this class you need a computer you can use at home. So, what kind is it (windows, Apple, Linux)? On a scale of 1 to 10, where 1 is a total dummy (doesn’t know a mouse from a keyboard) and 10 is a complete geek (knows computer stuff in and out) How do you rate yourself? You also need to have have internet access. How is your connection to the internet where you live?

1.2.1 my answer
I have a samsung chrome book laptop, It is a chome book.I would say it is a 5 in a scale of 1-10. Does not have a mouse but at home I integrated one in it. In a scale 1-10, I think that I am a 2 or 3 in skills with a computer. I have 1 gig of internet access.

1.3 Others and You?
What name do you prefer to be called by your classmates and instructor, if your buddys call you something but you want people in the class to call you something else - what would that be?

1.3.1 my answer
Name I prefer to to be called is Jesus. I actually really good with being called Jesus.

1.4 Aspirations?
What do you want to achieve in life? What college subject do you think would be of the greatest interest to you? What degree or certificate are you trying to achieve?

1.4.1 my answer
I want to be able to not worrie about money and be self sustained.I am currently studing to be RN.

1.5 Past adventures?
Have you ever gone to foreign countries, if so which ones, if not Where might you go in the future? Have you ever been in the military, if so which branch, if not which branch would you choose if you got drafted, and why?

1.5.1 my answer
I have been to Mexico in vacation to visit family. I want to be a traveling nurse so I can travel to other cities. I have not been to the military.

1.6 Why are you taking Statistics?
Did you take Math 115 or Math 116 ? Did you know Math 115 only satisfies a general education requirement but that Math 116 counts as taking statistics in several transfer institutions??
What grade do you need to attain in stats to not upset your future plans, What is your plan?

1.6.1 my answer
I am learning statistics to satisfy my prerequisite. I will be happy that this satisfies my general studies in math. I would like to at least receive a B or higher.